Fix authentication issues: hide navbar on auth page and protect routes
- Hide navigation bar on /auth page to prevent premature access to authenticated routes
- Create ProtectedRoute component for consistent authentication checking across pages
- Wrap all authenticated pages (dashboard, memories, life graph, how-to-use, api-docs, pro) with route protection
- Unauthenticated users are automatically redirected to /auth page
- Maintains existing auth flow while preventing unauthorized access
