+++ openmemory/ui/app/my-life/page.tsx
@@ -1,13 +1,12 @@
 "use client";
 
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import KnowledgeGraph from "./components/KnowledgeGraph";
 // import ChatInterface from "./components/ChatInterface";
 import { motion } from "framer-motion";
 import { Button } from "@/components/ui/button";
 import { MessageSquare, Network } from "lucide-react";
-import { useAuth } from "@/contexts/AuthContext";
-import { useRouter } from "next/navigation";
+import { ProtectedRoute } from "@/components/ProtectedRoute";
 /*
 import {
   Collapsible,
@@ -19,37 +18,11 @@ import {
 export default function MyLifePage() {
   const [selectedMemory, setSelectedMemory] = useState<string | null>(null);
   const [mobileView, setMobileView] = useState<"graph" | "chat">("graph");
-  const { user, isLoading } = useAuth();
-  const router = useRouter();
   const [isChatOpen, setIsChatOpen] = useState(true);
 
-  // Redirect unauthenticated users
-  useEffect(() => {
-    if (!isLoading && !user) {
-      router.push('/auth');
-    }
-  }, [user, isLoading, router]);
-
-  // Show loading while checking authentication
-  if (isLoading) {
-    return (
-      <div className="min-h-screen bg-zinc-950 flex items-center justify-center">
-        <div className="text-white">Loading...</div>
-      </div>
-    );
-  }
-
-  // Don't render if user is not authenticated
-  if (!user) {
-    return (
-      <div className="min-h-screen bg-zinc-950 flex items-center justify-center">
-        <div className="text-white">Redirecting to login...</div>
-      </div>
-    );
-  }
-
   return (
-    <div className="h-[calc(100vh-3.5rem)] lg:h-[calc(100vh-3.5rem)] flex flex-col lg:flex-row bg-background text-foreground">
+    <ProtectedRoute>
+      <div className="h-[calc(100vh-3.5rem)] lg:h-[calc(100vh-3.5rem)] flex flex-col lg:flex-row bg-background text-foreground">
       {/* Mobile Toggle */}
       <div className="lg:hidden flex items-center justify-center gap-2 p-2 bg-card border-b border-border">
         <Button
@@ -127,5 +100,6 @@ export default function MyLifePage() {
       </div>
       */}
     </div>
+    </ProtectedRoute>
   );
 }
